+++ src/algorithm/InsertionSort.cpp
@@ -1,21 +1,20 @@
 #include "algorithm.h"
 
-void insertionSort(std::vector<int> &arr, int &countComparison)
+void insertionSort(std::vector<int> &arr, unsigned long long &countComparison)
 {
     int n = arr.size();
-    for (int i = 1; ++countComparison && i < n; ++i) 
+    for (int i = 1; ++countComparison && i < n; ++i)
     {
         int key = arr[i];
         int j = i - 1;
 
-        while (++countComparison && j >= 0 && ++countComparison &&  arr[j] > key) 
+        while (++countComparison && j >= 0 && ++countComparison && arr[j] > key)
         {
-          
+
             arr[j + 1] = arr[j];
             j = j - 1;
         }
-        
+
         arr[j + 1] = key;
     }
 }
-
